% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln3.R
\docType{data}
\name{eln3}
\alias{eln3}
\title{ELN transcript variant 3}
\format{
A FASTA file with one sequence saved as a DNAString object
\describe{
  \item{first row}{header, a short description of the sequence contained in the file}
  \item{body}{the sequence of the transcript}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/nuccore/NM_001081753.3}
}
\usage{
eln3
}
\description{
A file containing the sequence of the ELN transctipt variant 3
}
\keyword{datasets}
