% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalSequenceAlignment.R
\name{globalSequenceAlignment}
\alias{globalSequenceAlignment}
\title{Sequence alignment}
\usage{
globalSequenceAlignment(seq1, seq2, gap = -1, miss = -1, match = 1)
}
\arguments{
\item{seq1}{A sequence as Biostrings::DNAString object}

\item{seq2}{A sequence as Biostrings::DNAString object}

\item{gap}{An integer representing the score for a gap in the alignment (default = -1)}

\item{miss}{An integer representing the score for a mismatch in the alignment (default = -1)}

\item{match}{An integer representing the score for a match in the alignment (default = +1)}
}
\value{
A list containing the final alignment, the score of the alignment and the score and directions matrices.
}
\description{
This function performs pairwise global sequence alignment of two sequences.
It uses the parameters gap, miss and match in order to compute the alignment
matrix and the directions matrix and then prints the resulting alignment
between the two sequences
}
